requirements.txt
openai>=1.0.0
anthropic>=0.30.0
google-generativeai>=0.8.0
datasets>=2.20.0
python-dotenv>=1.0.0
tqdm>=4.0.0
